refactor: extract date-time-picker core CSS to styles folder
